report.config file to avoid circular dependencies
