+++ apps/fishing-map/features/reports/reports.config.ts
@@ -1,0 +1,9 @@
+export type LastReportStorage = { reportUrl: string; workspaceUrl: string }
+export const LAST_REPORTS_STORAGE_KEY = 'lastReports'
+
+export const EMPTY_API_VALUES = ['NULL', undefined, '']
+export const MAX_CATEGORIES = 5
+export const OTHERS_CATEGORY_LABEL = 'OTHERS'
+
+export const MAX_DAYS_TO_COMPARE = 100
+export const MAX_MONTHS_TO_COMPARE = 12
