+++ apps/fishing-map/features/reports/reports.selectors.ts
@@ -29,11 +29,12 @@ import { selectContextAreasDataviews } from 'features/dataviews/dataviews.select
 import { createDeepEqualSelector } from 'utils/selectors'
 import { EMPTY_FIELD_PLACEHOLDER } from 'utils/info'
 import { sortStrings } from 'utils/shared'
+import {
+  EMPTY_API_VALUES,
+  MAX_CATEGORIES,
+  OTHERS_CATEGORY_LABEL,
+} from 'features/reports/reports.config'
 import { selectReportVesselsData } from './report.slice'
-
-export const EMPTY_API_VALUES = ['NULL', undefined, '']
-export const MAX_CATEGORIES = 5
-export const OTHERS_CATEGORY_LABEL = 'OTHERS'
 
 export type ReportVesselWithMeta = ReportVessel & {
   sourceColor: string
@@ -43,6 +44,7 @@ export type ReportVesselWithMeta = ReportVessel & {
   flagTranslated: string
   flagTranslatedClean: string
 }
+
 export type ReportVesselWithDatasets = Pick<ReportVessel, 'vesselId' | 'shipName' | 'hours'> &
   Partial<ReportVessel> &
   Pick<ReportVesselWithMeta, 'sourceColor'> & {
